requirements.txt
Flask==3.0.3
langchain==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
langchain-groq==0.2.0
PyPDF2==3.0.1
neo4j==5.24.0           
rdflib==7.0.0         
owlrl==6.0.2            
python-dotenv==1.0.1
pytest==8.3.2